script_print_ava/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pyautogui==0.9.54
Pillow==10.1.0

